refactoring to create generic 'solutions' template
