+++ components/solutions/comparisonRow.tsx
