+++ components/solutions/comparisonTable.tsx
