+++ components/solutions/index.tsx
