+++ components/solutions/sectionDivider.tsx
